# public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "agentsfleet Control Plane API",
    "description": "API for managing workspaces, fleets, triggers, and runs.",
    "contact": {
      "name": "agentsfleet",
      "url": "https://agentsfleet.net"
    },
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://api.agentsfleet.net",
      "description": "Production"
    }
  ],
  "paths": {
...
    "/v1/admin/fleet-libraries": {
      "get": {
        "tags": [
          "Fleet library"
        ],
        "summary": "List the platform Fleet library catalog",
        "description": "Lists every entry in the global platform catalog. Published, draft, and entries whose bundle was never fetched all appear. Unlike the workspace gallery, this operator view hides nothing: it shows what is live and what still needs work. Requires the `platform- library:write` scope. Metadata only — never bundle markdown, a support-file body, or an object-store key. Each row carries an `etag` that an editor can send as `If-Match` on PATCH. ",
        "operationId": "list_platform_fleet_library",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminLibrariesResponse"
                }
              }
            }
          },
          "401": {
            "description": "No credential, or one this route does not accept"
          },
          "403": {
            "description": "The credential is good and lacks the capability this route requires"
          },
          "500": {
            "description": "The daemon failed to answer"
          }
        }
      },
      "post": {
        "tags": [
          "Fleet library"
        ],
        "summary": "Onboard a platform Fleet library entry",
        "description": "Onboards a Fleet library entry into the global platform catalog from a GitHub source reference. Requires the `platform-library:write` scope. The canonical bundle is written to internal object storage keyed by content hash; the response carries metadata only — never an object-store key or support-file content. ",
        "operationId": "onboard_platform_fleet_library",
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "The request could not be read"
          },
          "401": {
            "description": "No credential, or one this route does not accept"
          },
          "403": {
            "description": "The credential is good and lacks the capability this route requires"
          },
          "409": {
            "description": "The request conflicts with the resource's current state"
          },
          "413": {
            "description": "The payload is over this route's ceiling"
          },
          "500": {
            "description": "The daemon failed to answer"
          }
        }
      }
    },
    "/v1/admin/fleet-libraries/{id}": {
      "delete": {
        "tags": [
          "Fleet library"
        ],
        "summary": "Delete an unpublished platform Fleet library entry",
        "description": "Removes a catalog entry. Only an entry that is NOT published may be deleted. A live fleet is never taken away from the tenants who can install it, so unpublish it first. Workspaces that already installed the fleet are unaffected: an install snapshots the bundle, so it keeps running. Requires the `platform-library:write` scope. ",
        "operationId": "delete_platform_fleet_library",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "`UUIDv7` of the resource.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No content"
          },
          "401": {
            "description": "No credential, or one this route does not accept"
          },
          "403": {
            "description": "The credential is good and lacks the capability this route requires"
          },
          "404": {
            "description": "No such resource, or none this caller may see"
          },
          "409": {
            "description": "The request conflicts with the resource's current state"
          },
          "500": {
            "description": "The daemon failed to answer"
          }
        }
      },
      "patch": {
        "tags": [
          "Fleet library"
        ],
        "summary": "Curate, publish, or unpublish a platform Fleet library entry",
        "description": "Partial update. `description` and `required_credentials_reasons` are the two fields no bundle can supply, so they are operator-owned: a later bundle refetch never overwrites them. `published` moves the entry between `draft` (stored, invisible to every tenant) and `public` (live in every workspace gallery and installable). Publishing an entry whose bundle was never fetched is refused — a published entry always has something to install. Requires the `platform-library:write` scope. Send `If-Match` with the row's `etag` to reject stale edits before they can repoint the source or unpublish the entry. Omitting the header preserves last-write-wins behavior. ",
        "operationId": "update_platform_fleet_library",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "`UUIDv7` of the resource.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "If-Match",
            "in": "header",
            "description": "Optional catalog row version from the list response. Stale values return 412 with the current `etag`.",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "The request could not be read"
          },
          "401": {
            "description": "No credential, or one this route does not accept"
          },
          "403": {
            "description": "The credential is good and lacks the capability this route requires"
          },
          "404": {
            "description": "No such resource, or none this caller may see"
          },
          "409": {
            "description": "The request conflicts with the resource's current state"
          },
          "412": {
            "description": "The precondition the caller supplied does not hold"
          },
          "500": {
            "description": "The daemon failed to answer"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AdminLibrariesResponse": {
        "type": "object",
        "description": "Admin Fleet-library list response.",
        "required": [
          "entries"
        ],
        "properties": {
          "entries": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AdminLibraryItem"
            },
            "description": "Every draft and public row."
          }
        },
        "additionalProperties": false
      },
      "AdminLibraryItem": {
        "type": "object",
        "description": "One metadata-only platform Fleet-library row.",
        "required": [
          "id",
          "name",
          "description",
          "source_repo",
          "source_ref",
          "visibility",
          "requirements",
          "required_credentials_reasons",
          "updated_at",
          "etag"
        ],
        "properties": {
          "content_hash": {
            "type": [
              "string",
              "null"
            ],
            "description": "Content identity, never support-file bytes."
          },
          "description": {
            "type": "string",
            "description": "Curated description."
          },
          "etag": {
            "type": "string",
            "description": "Strong version over the editable row surface."
          },
          "id": {
            "type": "string",
            "description": "Slug identity."
          },
          "name": {
            "type": "string",
            "description": "Display name."
          },
          "required_credentials_reasons": {
            "description": "Operator-authored per-credential reason copy."
          },
          "requirements": {
            "$ref": "#/components/schemas/AdminLibraryRequirements",
            "description": "Derived requirement names and trigger presence."
          },
          "source_ref": {
            "type": "string",
            "description": "Fetched revision."
          },
          "source_repo": {
            "type": "string",
            "description": "GitHub owner/repository."
          },
          "updated_at": {
            "type": "integer",
            "format": "int64",
            "description": "Last mutation instant in epoch milliseconds."
          },
          "visibility": {
            "type": "string",
            "description": "Draft or public."
          }
        },
        "additionalProperties": false
      },
      "AdminLibraryRequirements": {
        "type": "object",
        "description": "Content-free requirements shown on one Fleet-library row.",
        "required": [
          "credentials",
          "tools",
          "network_hosts",
          "trigger_present"
        ],
        "properties": {
          "credentials": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Credential names only."
          },
          "network_hosts": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Declared outbound hosts."
          },
          "tools": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Required tool names."
          },
          "trigger_present": {
            "type": "boolean",
            "description": "Whether a trigger document exists."
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Obtain a token via the CLI auth flow (POST /v1/auth/sessions) or GitHub OAuth"
      }
    }
  }
}